+++ src/main/java/com/service/point/dto/request/PointPolicyRegisterRequestDto.java
@@ -10,5 +10,5 @@ import lombok.Setter;
 public class PointPolicyRegisterRequestDto {
 
     private String pointAccumulationType;
-    private Integer pointValue;
+    private Long pointValue;
 }
